pyproject.toml
[project]
name = "framework"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "cryptography>=46.0.7",
    "fastapi>=0.136.1",
    "google-api-python-client>=2.194.0",
    "google-auth>=2.49.2",
    "google-auth-oauthlib>=1.3.1",
    "httpx>=0.28.1",
    "pydantic>=2.13.3",
    "pydantic-settings>=2.14.0",
    "python-dotenv>=1.2.2",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn[standard]>=0.46.0",
]

[dependency-groups]
dev = [
    "mypy>=1.20.2",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.11",
]
